requirements.txt
asgiref==3.6.0
asttokens==2.2.1
awsebcli==3.20.5
backcall==0.2.0
boto3==1.26.96
botocore==1.29.96
build==0.10.0
cement==2.8.2
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.6
Django==4.1.7
django-ckeditor==6.3.0
django-extensions==3.2.1
django-js-asset==2.0.0
django-storages==1.13.2
djangoajax==3.3
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.9.0
fonttools==4.38.0
idna==3.4
imageio==2.26.0
ipykernel==6.16.0
ipython==7.4.0
jedi==0.18.2
jmespath==1.0.1
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==1.4.4
lazy_loader==0.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
networkx==2.6
numpy==1.24.2
openpyxl==3.1.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathspec==0.10.1
pickleshare==0.7.5
Pillow==9.4.0
pip-tools==6.12.3
platformdirs==2.6.2
prompt-toolkit==2.0.10
psutil==5.9.4
psycopg2-binary==2.9.5
pure-eval==0.2.2
Pygments==2.14.0
pyparsing==3.0.9
pypiwin32==223
pyproject_hooks==1.0.0
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
pywin32==305
PyYAML==5.4.1
pyzmq==25.0.0
requests==2.26.0
s3transfer==0.6.0
scikit-image==0.20.0
scipy==1.10.1
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.3
stack-data==0.6.2
termcolor==1.1.0
tifffile==2023.2.28
tornado==6.2
traitlets==5.9.0
tzdata==2022.7
urllib3==1.26.15
virtualenv==20.17.1
wcwidth==0.1.9
